+++ src/open_modal.py
@@ -251,65 +251,3 @@ def open_project_modal(current_project):
         is_open=True,
     )
     return modal
-
-
-def add_dataset_cb(
-        new_data, project_name_checked, dataset_name,
-        text_column, label_column, description,
-        project_name, current_dataset):
-    """Callback for the add dataset button.
-
-    The function assumes everything to be valid, since it can only be trigger if
-    the validator is true.
-    """
-    datasets.create_dataset(new_data, dataset_name, description, text_column)
-    if project_name_checked:
-        new_current_project = datasets.create_project(dataset_name, project_name, label_column)
-        return False, {
-            'dataset_name': dataset_name,
-            'project_name': project_name,
-            'data': new_current_project.to_dict('records')
-        }
-    else:
-        return False, current_dataset
-
-
-def create_project_cb(
-        create_project_name_valid, create_proj_dd_selection,
-        create_project_name, current_dataset, label_column):
-    """callback for create Project Button.
-
-    Args:
-        create_project_name_valid: whether the new project name is valid
-        create_proj_dd_selection: selected dataset for project creation
-        create_project_name: name for new project
-        current_dataset: currently loaded datasets
-        n_dataset_loads: number of datasets loads so far
-
-    Returns:
-        The info about whether the Modal should
-        be closed and the updated current dataset
-    """
-    if create_project_name_valid and create_proj_dd_selection:
-        new_current_project = datasets.create_project(
-            create_proj_dd_selection,
-            create_project_name,
-            label_column
-        )
-        return False, {
-            'dataset_name': create_proj_dd_selection,
-            'project_name': create_project_name,
-            'data': new_current_project.to_dict('records')
-        }
-    else:
-        return True, current_dataset
-
-
-def open_project_cb(
-        open_project_dd_selection):
-    new_current_project, dataset_name = datasets.load_project(open_project_dd_selection)
-    return False, {
-        'dataset_name': dataset_name,
-        'project_name': open_project_dd_selection,
-        'data': new_current_project.to_dict('records')
-    }
